Agno_with_mercury/requirements.txt
agno==2.5.10
ddgs==9.11.4
openai==2.29.0
python-dotenv==1.2.2
yfinance==1.2.0